Code split between local/browser specific code for slightly faster startup/loading
