+++ newIDE/app/src/index.js
@@ -1,11 +1,7 @@
 // @flow
 import 'element-closest';
-import React from 'react';
+import React, { Component, type Element } from 'react';
 import ReactDOM from 'react-dom';
-import MainFrame from './MainFrame';
-import Window from './Utils/Window';
-import ExportDialog from './Export/ExportDialog';
-import CreateProjectDialog from './ProjectCreation/CreateProjectDialog';
 import Authentification from './Utils/GDevelopServices/Authentification';
 import {
   sendProgramOpening,
@@ -15,32 +11,8 @@ import { installRaven } from './Utils/Analytics/Raven';
 import { installFullstory } from './Utils/Analytics/Fullstory';
 import { unregister } from './registerServiceWorker';
 import './UI/iconmoon-font.css'; // Styles for Iconmoon font.
-
-// Import for browser only IDE
-import BrowserExamples from './ProjectCreation/BrowserExamples';
-import BrowserStarters from './ProjectCreation/BrowserStarters';
-import BrowserProjectOpener from './ProjectsStorage/BrowserProjectOpener';
-import BrowserSaveDialog from './ProjectsStorage/BrowserSaveDialog';
-import BrowserIntroDialog from './MainFrame/BrowserIntroDialog';
-import browserResourceSources from './ResourcesList/BrowserResourceSources';
-import browserResourceExternalEditors from './ResourcesList/BrowserResourceExternalEditors';
-import BrowserS3PreviewLauncher from './Export/BrowserExporters/BrowserS3PreviewLauncher';
-import { getBrowserExporters } from './Export/BrowserExporters';
-import BrowserJsExtensionsLoader from './JsExtensionsLoader/BrowserJsExtensionsLoader';
-
-// Import for Electron powered IDE.
-import ExternalEditor from './ExternalEditor';
 import optionalRequire from './Utils/OptionalRequire.js';
-import LocalExamples from './ProjectCreation/LocalExamples';
-import LocalStarters from './ProjectCreation/LocalStarters';
-import localResourceSources from './ResourcesList/LocalResourceSources';
-import localResourceExternalEditors from './ResourcesList/LocalResourceExternalEditors';
-import LocalProjectWriter from './ProjectsStorage/LocalProjectWriter';
-import LocalProjectOpener from './ProjectsStorage/LocalProjectOpener';
-import LocalPreviewLauncher from './Export/LocalExporters/LocalPreviewLauncher';
-import { getLocalExporters } from './Export/LocalExporters';
-import ElectronEventsBridge from './MainFrame/ElectronEventsBridge';
-import LocalJsExtensionsLoader from './JsExtensionsLoader/LocalJsExtensionsLoader';
+import { showErrorBox } from './UI/Messages/MessageBox';
 
 // Uncomment to enable logs in console when a component is potentially doing
 // an unnecessary update
@@ -49,83 +21,57 @@ import LocalJsExtensionsLoader from './JsExtensionsLoader/LocalJsExtensionsLoade
 
 const electron = optionalRequire('electron');
 
-const authentification = new Authentification();
-installAnalyticsEvents(authentification);
-installRaven();
-installFullstory();
+type State = {|
+  App: ?Element<*>,
+|};
 
-Window.setUpContextMenu();
+class Bootstrapper extends Component<{}, State> {
+  state = {
+    App: null,
+  };
+  authentification = new Authentification();
 
-let app = null;
-const appArguments = Window.getArguments();
+  componentDidMount() {
+    installAnalyticsEvents(this.authentification);
+    installRaven();
+    installFullstory();
 
-if (electron) {
-  if (appArguments['server-port']) {
-    app = (
-      <ExternalEditor
-        serverPort={appArguments['server-port']}
-        isIntegrated={appArguments['mode'] === 'integrated'}
-        editor={appArguments['editor']}
-        editedElementName={appArguments['edited-element-name']}
-      >
-        <MainFrame
-          resourceSources={localResourceSources}
-          authentification={authentification}
-          onReadFromPathOrURL={() => Promise.reject('Should never be called')}
-          resourceExternalEditors={localResourceExternalEditors}
-          initialPathsOrURLsToOpen={[]}
-        />
-      </ExternalEditor>
-    );
-  } else {
-    app = (
-      <ElectronEventsBridge>
-        <MainFrame
-          previewLauncher={<LocalPreviewLauncher />}
-          exportDialog={<ExportDialog exporters={getLocalExporters()} />}
-          createDialog={
-            <CreateProjectDialog
-              examplesComponent={LocalExamples}
-              startersComponent={LocalStarters}
-            />
-          }
-          onSaveProject={LocalProjectWriter.saveProject}
-          onChooseProject={LocalProjectOpener.chooseProjectFile}
-          onReadFromPathOrURL={LocalProjectOpener.readProjectJSONFile}
-          resourceSources={localResourceSources}
-          resourceExternalEditors={localResourceExternalEditors}
-          authentification={authentification}
-          extensionsLoader={new LocalJsExtensionsLoader()}
-          initialPathsOrURLsToOpen={appArguments['_']}
-        />
-      </ElectronEventsBridge>
-    ); 
+    if (electron) {
+      import(/* webpackChunkName: "local-app" */ './LocalApp')
+        .then(module =>
+          this.setState({
+            App: module.create(this.authentification),
+          })
+        )
+        .catch(this.handleLoadError);
+    } else {
+      import(/* webpackChunkName: "browser-app" */ './BrowserApp')
+        .then(module =>
+          this.setState({
+            App: module.create(this.authentification),
+          })
+        )
+        .catch(this.handleLoadError);
+    }
   }
-} else {
-  app = (
-    <MainFrame
-      previewLauncher={<BrowserS3PreviewLauncher />}
-      exportDialog={<ExportDialog exporters={getBrowserExporters()} />}
-      createDialog={
-        <CreateProjectDialog
-          examplesComponent={BrowserExamples}
-          startersComponent={BrowserStarters}
-        />
-      }
-      introDialog={<BrowserIntroDialog />}
-      saveDialog={<BrowserSaveDialog />}
-      onReadFromPathOrURL={BrowserProjectOpener.readInternalFile}
-      resourceSources={browserResourceSources}
-      resourceExternalEditors={browserResourceExternalEditors}
-      authentification={authentification}
-      extensionsLoader={new BrowserJsExtensionsLoader()}
-      initialPathsOrURLsToOpen={appArguments['_']}
-    />
-  );
+
+  handleLoadError(err) {
+    const message = !electron
+      ? 'Please reload the page and check your internet connectivity.'
+      : 'Please restart the application or reinstall the latest version if the problem persists.';
+    showErrorBox(`Unable to load GDevelop. ${message}`, err);
+  }
+
+  render() {
+    const { App } = this.state;
+
+    if (App) return App;
+    return null;
+  }
 }
 
 const rootElement = document.getElementById('root');
-if (rootElement) ReactDOM.render(app, rootElement);
+if (rootElement) ReactDOM.render(<Bootstrapper />, rootElement);
 else console.error('No root element defined in index.html');
 
 // registerServiceWorker();
